+++ src/main/core/sendAudioData.ts
@@ -1,75 +1,14 @@
 import { addSongToPlayHistory } from '@main/db/queries/history';
 import { getPlayableSongById } from '@main/db/queries/songs';
 import { setDiscordRpcActivity } from '@main/other/discordRPC';
-import sharp from 'sharp';
-
 import {
   parseArtistOnlineArtworks,
   parseSongArtworks,
-  removeDefaultAppProtocolFromFilePath,
   resolveSongFilePath
 } from '../fs/resolveFilePaths';
 import logger from '../logger';
-import { IS_DEVELOPMENT, setCurrentSongPath } from '../main';
+import { setCurrentSongPath } from '../main';
 import { parsePaletteFromArtworks } from './getAllSongs';
-
-export const parseArtworkDataForAudioPlayerData = (artworkData?: Buffer | Uint8Array) => {
-  if (artworkData === undefined) return undefined;
-
-  if (IS_DEVELOPMENT) return Buffer.from(artworkData).toString('base64');
-  return artworkData;
-};
-
-// const getRelevantArtistData = (
-//   songArtists?: {
-//     artistId: string;
-//     name: string;
-//   }[]
-// ) => {
-//   const artists = getArtistsData();
-//   const relevantArtists: {
-//     artistId: string;
-//     artworkName?: string;
-//     name: string;
-//     onlineArtworkPaths?: OnlineArtistArtworks;
-//   }[] = [];
-
-//   if (songArtists) {
-//     for (const songArtist of songArtists) {
-//       for (const artist of artists) {
-//         if (artist.artistId === songArtist.artistId) {
-//           if (!artist.onlineArtworkPaths)
-//             getArtistInfoFromNet(artist.artistId).catch((error) =>
-//               logger.warn('Failed to get artist info from net', { err: error })
-//             );
-
-//           const { artistId, name, artworkName, onlineArtworkPaths } = artist;
-
-//           relevantArtists.push({
-//             artistId,
-//             name,
-//             artworkName,
-//             onlineArtworkPaths
-//           });
-//         }
-//       }
-//     }
-//   }
-
-//   return relevantArtists;
-// };
-
-const getArtworkBuffer = async (artworkPath: string) => {
-  try {
-    const realPath = removeDefaultAppProtocolFromFilePath(artworkPath);
-    const buffer = await sharp(realPath).toBuffer();
-
-    return buffer;
-  } catch {
-    // Failed to get artwork buffer most probably becuase the artwork path is a packaged path
-    return undefined;
-  }
-};
 
 const sendAudioData = async (
   songId: number,
@@ -90,7 +29,6 @@ const sendAudioData = async (
       const artworks = song.artworks.map((a) => a.artwork);
       const artworkPaths = parseSongArtworks(artworks);
       const songArtwork = artworkPaths.artworkPath;
-      const artworkData = await getArtworkBuffer(songArtwork);
 
       const albumObj = song.albums?.[0]?.album;
       const album = albumObj ? { albumId: albumObj.id, name: albumObj.title } : undefined;
@@ -101,7 +39,6 @@ const sendAudioData = async (
         title: song.title,
         artists,
         duration: Number(song.duration),
-        artwork: parseArtworkDataForAudioPlayerData(artworkData),
         artworkPath: songArtwork,
         path: resolveSongFilePath(song.path),
         songId: song.id,